% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{app_old}
\alias{app_old}
\title{Internal Shiny App for CPAM Visualization}
\usage{
app_old(cpo)
}
\arguments{
\item{cpo}{A CPAM object containing fitted models and expression data}
}
\value{
A Shiny app object
}
\description{
Internal Shiny App for CPAM Visualization
}
\keyword{internal}
