% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed-effects.R
\name{check_collinearity}
\alias{check_collinearity}
\title{Check for collinearity between predictors and with time variable}
\usage{
check_collinearity(exp_design, formula)
}
\arguments{
\item{exp_design}{Experimental design data frame}

\item{formula}{Model formula}
}
\value{
Logical indicating whether collinearity was detected
}
\description{
Check for collinearity between predictors and with time variable
}
\keyword{internal}
