% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed-effects.R
\name{check_collinearity_with_time}
\alias{check_collinearity_with_time}
\title{Check for collinearity between fixed effects and time variable}
\usage{
check_collinearity_with_time(exp_design, fe_terms)
}
\arguments{
\item{exp_design}{Experimental design data frame}

\item{fe_terms}{Fixed effects terms to check}
}
\value{
Logical indicating whether collinearity was detected
}
\description{
Check for collinearity between fixed effects and time variable
}
