% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_cpam.R
\name{compute_normalization_factors}
\alias{compute_normalization_factors}
\title{Compute normalization factors}
\usage{
compute_normalization_factors(count_matrix_filtered, normalize)
}
\arguments{
\item{count_matrix_filtered}{Filtered count matrix}

\item{normalize}{Whether to perform normalization}
}
\value{
Vector of normalization factors
}
\description{
Compute normalization factors
}
\keyword{internal}
