% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_p_values.R
\name{compute_p_values}
\alias{compute_p_values}
\title{Compute P values}
\usage{
compute_p_values(
  cpo,
  subset = NULL,
  p_adj_method = "BH",
  fixed_effects = NULL,
  gam_method = "REML",
  gam_optimizer = "efs",
  silent = TRUE
)
}
\arguments{
\item{cpo}{a cpam object}

\item{subset}{a character vector of target_id names}

\item{p_adj_method}{method for P value adjustment}

\item{fixed_effects}{a model formula (RHS only) to provide or update
fixed affects. Terms must correspond to columns in \code{exp_design}}

\item{gam_method}{fitting method for \code{mgcv::gam}}

\item{gam_optimizer}{optimization method for \code{mgcv::gam}}

\item{silent}{logical; silences warnings from model fitting}
}
\value{
an updated cpam object with P values stored in the new slot "p_table"
}
\description{
Compute P values
}
\examples{
1 + 1
}
