% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_cpam.R
\name{convert_to_long_format}
\alias{convert_to_long_format}
\title{Convert count matrix to long format and join with experiment design}
\usage{
convert_to_long_format(counts_raw, exp_design)
}
\arguments{
\item{counts_raw}{Raw count matrix}

\item{exp_design}{Experimental design}
}
\value{
Long format data frame
}
\description{
Convert count matrix to long format and join with experiment design
}
