% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{count_matrix_example}
\alias{count_matrix_example}
\title{A small case-only count matrix to run light-weight examples}
\format{
A matrix with 1000 rows and 30 columns:
\describe{
\item{rownames}{Gene ID}
\item{colnames}{Samples}
}
}
\source{
Simulated data, see data-raw directory for details
}
\usage{
count_matrix_example
}
\description{
See \code{cpo_example} for fitted CPAM object.
See \code{exp_design_example} for experimental design.
1000 genes x 6 time points x 5 reps
}
\keyword{datasets}
