% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpam-package.R
\docType{package}
\name{cpam-package}
\alias{cpam}
\alias{cpam-package}
\title{cpam: Changepoint additive models for time series omics data}
\description{
cpam provides a comprehensive framework for time series omics analysis, integrating changepoint detection, smooth and shape-constrained trends, and uncertainty quantification. It supports gene- and transcript-level inferences, p-value aggregation for improved power, and both case-only and case-control designs. It includes an interactive Shiny interface.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://l-a-yates.github.io/cpam/}
}

}
\author{
\strong{Maintainer}: Luke Yates \email{luke.yates@utas.edu.au} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
