% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cpo_example}
\alias{cpo_example}
\title{A fitted \code{cpam} object to run light-weight examples}
\format{
A cpam object with p-values, changpoints, and shapes estimated
}
\source{
Simulated data, see data-raw directory for details
}
\usage{
cpo_example
}
\description{
Six time points x 3 reps
See \code{cpo_example} for fitted CPAM object.
See \code{exp_design_example} for experimental design.
#'
}
\concept{dataset}
\keyword{datasets}
