% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_changepoint.R
\name{dnbl}
\alias{dnbl}
\title{Calculate log probability density for negative binomial distribution}
\usage{
dnbl(x, theta, mu)
}
\arguments{
\item{x}{Response values}

\item{theta}{Dispersion parameter}

\item{mu}{Mean parameter}
}
\value{
Log probability density values
}
\description{
Calculate log probability density for negative binomial distribution
}
\details{
Computes the log probability density for a negative binomial distribution
adapted for continuous positive response values.
}
\keyword{internal}
