% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_changepoint.R
\name{estimate_changepoint}
\alias{estimate_changepoint}
\title{Use model selection to estimate changepoints}
\usage{
estimate_changepoint(
  cpo,
  cps = NULL,
  degs_only = T,
  deg_threshold = 0.05,
  subset = NULL,
  sp = NULL,
  bss = "tp",
  family = c("nb", "gaussian"),
  score = "aic"
)
}
\arguments{
\item{cpo}{a cpam object}

\item{cps}{vector of candidate changepoints. Defaults to the set of observed timepoints}

\item{degs_only}{logical; should changepoints only be estimated for differentially expressed genes}

\item{deg_threshold}{logical; the threshold value for DEGs (ignored of \code{degs_only = F})}

\item{subset}{character vector; names of targets or genes (if \code{cpo$gene_level = T})
for which changepoints will be estimated}

\item{sp}{numerical >= 0; supply a fixed smoothing parameter.}

\item{bss}{character vector; names of candidate spline bases (i.e., candidate shape types).
Default is thin plate ("tp") splines.}

\item{family}{character; negative binomial ("nb", default) or Gaussian ("gaussian")}

\item{score}{character; model selection score, either Generalised Cross Validation ("gcv") or
Akaike Information Criterion ("aic")}
}
\value{
a cpam object with the estimated changepoint table added to the slot "changepoints"
}
\description{
Use model selection to estimate changepoints
}
\examples{
1+1
}
