% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_cpam.R
\name{estimate_dispersions_wrapper}
\alias{estimate_dispersions_wrapper}
\title{Wrapper for estimating dispersions}
\usage{
estimate_dispersions_wrapper(
  count_matrix_filtered,
  exp_design,
  bootstrap,
  catch
)
}
\arguments{
\item{count_matrix_filtered}{Filtered count matrix}

\item{exp_design}{Experimental design}

\item{bootstrap}{Whether bootstrap was used}

\item{catch}{Overdispersion calculations from bootstrap}
}
\value{
Estimated dispersions
}
\description{
Wrapper for estimating dispersions
}
