% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exp_design}
\alias{exp_design}
\title{Example Experimental Design}
\format{
A tibble with columns:
\describe{
\item{sample}{Character vector of sample names.}
\item{time}{Numeric vector of time points.}
\item{path}{Character vector of file paths.}
...
}
}
\source{
Simulated example.
}
\usage{
exp_design
}
\description{
A tibble representing an example experimental design for CPAM analysis.
}
\keyword{datasets}
