% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exp_design_example}
\alias{exp_design_example}
\title{An experimental design to run light-weight examples}
\format{
A tibble with columns:
\describe{
\item{sample}{Character vector of sample names.}
\item{time}{Numeric vector of time points.}
...
}
}
\source{
Simulated data, see data-raw directory for details
}
\usage{
exp_design_example
}
\description{
Six time points x 3 reps
#' See \code{cpo_example} for fitted CPAM object.
}
\details{
#'
}
\concept{dataset}
\keyword{datasets}
