% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_cpam.R
\name{filter_and_update_data}
\alias{filter_and_update_data}
\title{Filter and update data_long with normalization factors}
\usage{
filter_and_update_data(data_long, target_to_keep, norm_factor)
}
\arguments{
\item{data_long}{Long format data}

\item{target_to_keep}{Targets to keep after filtering}

\item{norm_factor}{Normalization factors}
}
\value{
Filtered and updated data
}
\description{
Filter and update data_long with normalization factors
}
\keyword{internal}
