% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_cpam.R
\name{handle_gene_aggregation}
\alias{handle_gene_aggregation}
\title{Handle gene aggregation in long format data}
\usage{
handle_gene_aggregation(data_long, t2g, aggregate_to_gene)
}
\arguments{
\item{data_long}{Long format data}

\item{t2g}{Transcript to gene mapping}

\item{aggregate_to_gene}{Whether to aggregate to gene level}
}
\value{
Updated long format data with gene ID information
}
\description{
Handle gene aggregation in long format data
}
\keyword{internal}
