% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_cpam.R
\name{import_count_data}
\alias{import_count_data}
\title{Import count data from files}
\usage{
import_count_data(exp_design, t2g, import_type, gene_level, bootstrap)
}
\arguments{
\item{exp_design}{Experimental design}

\item{t2g}{Transcript to gene mapping}

\item{import_type}{Import type}

\item{gene_level}{Whether to aggregate to gene level}

\item{bootstrap}{Whether to use bootstrap samples}
}
\value{
List containing imported data and related variables
}
\description{
Import count data from files
}
