% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-gene.R
\name{plot_cluster}
\alias{plot_cluster}
\title{Plot clustered targets}
\usage{
plot_cluster(cpo, res, changepoints, shapes, alpha = 0.1)
}
\arguments{
\item{cpo}{a cpam object}

\item{res}{a tibble, output from \code{\link[=results]{results()}}}

\item{changepoints}{numerical or character; one or more changepoints}

\item{shapes}{character; one or more shapes}

\item{alpha}{numerical; line transparency}
}
\value{
a ggplot object
}
\description{
Plot clustered targets
}
\examples{
1 + 1
}
