% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-gene.R
\name{plot_gene_co}
\alias{plot_gene_co}
\title{Plot a fitted cpgam}
\usage{
plot_gene_co(
  cpo,
  gene_id = NULL,
  target_id = NULL,
  cp_type = c("cp_min", "cp1se"),
  shape_type = "shape1",
  bs = "auto",
  cp_fix = -1,
  facet = F,
  sp = NULL,
  show_fit = T,
  show_data = T,
  show_fit_ci = T,
  show_data_se = T,
  ci_prob = 2 * pnorm(1) - 1,
  remove_null = F,
  gene_level = F,
  family = "nb"
)
}
\arguments{
\item{cpo}{a cpam object}

\item{gene_id}{character; gene_id}

\item{target_id}{character; target_id}

\item{cp_type}{character; if changepoints have been estimated using \code{\link[=estimate_changepoint]{estimate_changepoint()}},
which selection rule should be used. See \code{\link[=estimate_changepoint]{estimate_changepoint()}} for details.}

\item{shape_type}{character; if shapes have been estimated using \code{\link[=select_shape]{select_shape()}},
which set of candidate shapes should be used. See \code{\link[=select_shape]{select_shape()}} for details.}

\item{bs}{character; set the basis (i.e. shape)}

\item{cp_fix}{numerical; set the changepoint}

\item{facet}{logical; for genes with multiple transcripts, should the transcripts be plotted in separate facets}

\item{sp}{numerical; set the smooth parameter}

\item{show_fit}{logical; show the fitted trend}

\item{show_data}{logical; show (possibly normalized and scaled) data points}

\item{show_fit_ci}{logical; show credible interval for the fitted trend}

\item{show_data_se}{logical; show bootstrapped quantile for data points}

\item{ci_prob}{numerical; set the probabilty for the credible interval of the fitted trend}

\item{remove_null}{logical; only plot differentially expressed transcripts}

\item{gene_level}{logical; plot gene-level data and fitted trend}

\item{family}{character; negative binomial ("nb", default) or Gaussian ("gaussian")}
}
\value{
a ggplot object
}
\description{
Plot a fitted cpgam
}
\examples{
1+1
}
