% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_cpam.R
\name{prepare_case_control}
\alias{prepare_case_control}
\title{Prepare case-control data by adding case indicator}
\usage{
prepare_case_control(exp_design, condition_var, case_value)
}
\arguments{
\item{exp_design}{Experimental design data frame}

\item{condition_var}{Name of condition variable}

\item{case_value}{Value indicating case}
}
\value{
Updated experimental design with case indicator
}
\description{
Prepare case-control data by adding case indicator
}
