% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_cpam.R
\name{prepare_cpam}
\alias{prepare_cpam}
\title{Prepare a cpam object}
\usage{
prepare_cpam(
  exp_design,
  count_matrix = NULL,
  t2g = NULL,
  import_type = NULL,
  model_type = c("case-only", "case-control"),
  bootstrap = TRUE,
  nboot_max = NULL,
  filter_fun = "ts_filter",
  filter_fun_args = list(min_reads = 5, min_prop = 3/5),
  regularize = TRUE,
  gene_level = FALSE,
  aggregate_to_gene = !gene_level,
  condition_var = "condition",
  case_value = "treatment",
  num_cores = 1,
  normalize = TRUE,
  fixed_effects = NULL
)
}
\arguments{
\item{exp_design}{a dataframe or tibble with the experimental design, containing at least a 'time' and a 'sample' column}

\item{count_matrix}{a matrix of counts. Column names must be in 'sample' column of \code{exp_design},}

\item{t2g}{a transcript to gene dataframe or tibble with columns target_id and gene_id}

\item{import_type}{software used for quantification, one of "kallisto", "sailfish" ,...}

\item{model_type}{"case-only" (default) or "case-control"}

\item{bootstrap}{logical; load bootstrap samples, also called inferential replicates, if available, and rescale counts.}

\item{nboot_max}{Maximum number of bootstrap samples to use}

\item{filter_fun}{filter function to remove lowly expressed genes (default is \code{filter_fun()})}

\item{filter_fun_args}{arguments for filter function}

\item{regularize}{logical; use empirical Bayes regularization of dispersions (default is TRUE)}

\item{gene_level}{logical; aggregate counts to gene level before data preparation and modelling (default is FALSE)}

\item{aggregate_to_gene}{logical; aggregate P values from transcript- to gene-level}

\item{condition_var}{string; column name in \code{exp_design} for the condition variable (for \code{model_type} = "case_control" only)}

\item{case_value}{value of \code{condition_var} that indicates the "case". All other values are deemed to be control}

\item{num_cores}{integer; number of cores to use for parallel computation}

\item{normalize}{logical; use model offsets based on sampling depth and gene length}

\item{fixed_effects}{a model formula of the form \code{~ effect1 + effect2}}
}
\value{
an object of class \code{cpo}
}
\description{
Prepare a cpam object
}
\examples{
1+1
}
