% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_cpam.R
\name{process_bootstrap_data}
\alias{process_bootstrap_data}
\title{Process bootstrap data}
\usage{
process_bootstrap_data(data_long, catch, boot)
}
\arguments{
\item{data_long}{Long format data}

\item{catch}{Overdispersion calculations}

\item{boot}{Bootstrap summary}
}
\value{
Updated data with bootstrap information
}
\description{
Process bootstrap data
}
