% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_cpam.R
\name{process_count_matrix}
\alias{process_count_matrix}
\title{Process count matrix when importing is not needed}
\usage{
process_count_matrix(count_matrix, t2g, gene_level)
}
\arguments{
\item{count_matrix}{Count matrix}

\item{t2g}{Transcript to gene mapping}

\item{gene_level}{Whether to aggregate to gene level}
}
\value{
Processed count matrix
}
\description{
Process count matrix when importing is not needed
}
