% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_cpam.R
\name{summarise_bootstraps}
\alias{summarise_bootstraps}
\title{Summarize bootstrap samples}
\usage{
summarise_bootstraps(txi)
}
\arguments{
\item{txi}{A tximport object containing inferential replicates}
}
\value{
A tibble with bootstrap summary statistics
}
\description{
Internal function to process inferential replicates from tximport objects.
}
