% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{t2g}
\alias{t2g}
\title{Transcript-to-Gene Mapping}
\format{
A tibble with columns:
\describe{
\item{transcript}{Character vector of transcript IDs.}
\item{gene}{Character vector of corresponding gene IDs.}
}
}
\source{
The TAIR database.
}
\usage{
t2g
}
\description{
A tibble representing a mapping of transcript IDs to gene IDs.
}
\keyword{datasets}
