% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_cpam.R
\name{ts_filter}
\alias{ts_filter}
\title{Removes lowly expressed genes}
\usage{
ts_filter(data, min_reads = 5, min_prop = 3/5)
}
\arguments{
\item{data}{count data in long form}

\item{min_reads}{minimum reads per transcript per sample}

\item{min_prop}{minimum proportion of samples that exceed \code{min_read} at a given time point}
}
\value{
a character vector transcript id to keep
}
\description{
Removes lowly expressed genes
}
\examples{
1+1
}
