% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_cpam.R
\name{validate_cores}
\alias{validate_cores}
\title{Validate and adjust number of cores}
\usage{
validate_cores(num_cores)
}
\arguments{
\item{num_cores}{Number of cores requested}
}
\value{
Validated number of cores
}
\description{
Validate and adjust number of cores
}
