% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpam-package.R
\docType{package}
\name{cpam-package}
\alias{cpam}
\alias{cpam-package}
\title{cpam: Changepoint additive models for time series omics data}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Luke Yates \email{luke.yates@utas.edu.au} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
