% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{results}
\alias{results}
\title{Title}
\usage{
results(
  cpo,
  p_threshold = 0.05,
  p_type = c("p_gam", "p_mvn"),
  min_lfc = 0,
  min_count = 0,
  aggregate_to_gene = cpo$aggregate_to_gene,
  add_lfc = T,
  add_counts = T,
  cp_type = c("cp_1se", "cp_min"),
  shape_type = c("shape1", "shape2"),
  summarise_to_gene = F
)
}
\arguments{
\item{cpo}{a cpam object}

\item{p_threshold}{numerical; threshold}

\item{p_type}{character; choose the type of p-value}

\item{min_lfc}{numerical; maximum absolute log (base 2) fold change must exceed this minimum value; default is 0}

\item{min_count}{numerical; maximum of the modelled counts evaluated at the set of observed time points must exceed this minimum value for}

\item{aggregate_to_gene}{logical; use aggregate p-values}

\item{add_lfc}{logical; add log (base 2) fold changes for each time point}

\item{add_counts}{logical; add modelled counts for each time point}

\item{cp_type}{character; model-selection rule used to select the changepoint}

\item{shape_type}{character; "shape1" to include unconstrained or otherwise "shape2"}

\item{summarise_to_gene}{logical; return gene-level results only}
}
\value{
a tibble
}
\description{
Title
}
\examples{
1+1
}
