% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_shape.R
\name{select_shape}
\alias{select_shape}
\title{Use model selection to select a shape from a list of candidates}
\usage{
select_shape(
  cpo,
  subset = NULL,
  sp = NULL,
  bss = c("micv", "mdcx", "cv", "cx", "lin", "tp", "null"),
  family = c("nb", "gaussian"),
  score = "gcv",
  cp_type = c("cp_1se", "cp_min")
)
}
\arguments{
\item{cpo}{a cpam object}

\item{subset}{character vector; names of targets or genes (if \code{cpo$gene_level = T})
for which changepoints will be estimated}

\item{sp}{numerical >= 0; supply a fixed smoothing parameter.
Note, the fixd value is applied to shape constrained bases only (i.e., not \code{bs = 'tp'}).}

\item{bss}{character vector; names of candidate spline bases (i.e., candidate shape types).}

\item{family}{character; negative binomial ("nb", default) or Gaussian ("gaussian")}

\item{score}{character; model selection score, either Generalised Cross Validation ("gcv") or
Akaike Information Criterion ("aic")}

\item{cp_type}{character; if changepoints have been estimated using \code{\link[=estimate_changepoint]{estimate_changepoint()}},
which selection rule should be used. See \code{\link[=estimate_changepoint]{estimate_changepoint()}} for details.}
}
\value{
a cpam object with the selected shapes added to the slot "shapes"
}
\description{
Use model selection to select a shape from a list of candidates
}
\examples{
1+1
}
