% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{visualize}
\alias{visualize}
\title{Launches a Shiny app to visualise the data and fitted models of a cpam object}
\usage{
visualize(cpo, subset = NULL, degs_only = T, deg_threshold = 0.05)
}
\arguments{
\item{cpo}{a cpam object}

\item{subset}{character vector; names of targets or genes (if \code{cpo$gene_level = T})
to load into Shiny app}

\item{degs_only}{logical; display DEGs only}

\item{deg_threshold}{numerical; threshold for DEGs}
}
\value{
None
}
\description{
Launches a Shiny app to visualise the data and fitted models of a cpam object
}
\examples{
1+1
}
